refactor(test): share the browser-project derivation between its two guards
ci-verify-coverage.test.ts duplicated the vitest.config.ts browser-project
derivation already inline in docs-contract.test.ts. Extract it to
shared/test-utils/vitest-browser-projects.ts per the repo's shared-test-helper
convention; both guards keep asserting their own surface (docs enumeration,
test:browser script) against the same ground truth.
